[ #10 ] Updated type for executionId variable from Integer to Long to avoid integer overflow.

+++ modules/yaf-zephyr-scale/src/main/java/com/coherentsolutions/yaf/zephyr/scale/domain/CycleCreateDTO.java
@@ -45,5 +45,5 @@ public class CycleCreateDTO extends Cycle {
     private int projectVersion;
     private String projectKey;
     private String name;
-    private Integer folderId;
+    private Long folderId;
 }
